+++ example/lib/responsive_collection/responsive_padding.dart
@@ -1,8 +1,6 @@
 import 'package:responsive_spacing/responsive_spacing.dart';
 
 class MyResponsivePadding extends ResponsiveCollection {
-  @override
-  ScaledSize fallback(double width) => const ScaledSize(size: 8);
 
   @override
   ScaledSize xl(double width) => const ScaledSize(size: 32);