+++ lib/widgets/responsive_card.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:responsive_spacing/extensions/scaled_size_extension.dart';
 import 'package:responsive_spacing/widgets/spacing.dart';
 
-/// It is basically a Card
+/// It is basically a Card with a padding
 class ResponsiveCard extends Card {
   /// This card adds padding around the Card to fit the sizes
   const ResponsiveCard({
-    Key? key,
+    super.key,
     super.color,
     super.shadowColor,
     super.surfaceTintColor,
@@ -16,7 +16,7 @@ class ResponsiveCard extends Card {
     super.clipBehavior,
     super.child,
     super.semanticContainer = true,
-  }) : super(margin: EdgeInsets.zero, key: key);
+  }) : super(margin: EdgeInsets.zero);
 
   @override
   Widget build(BuildContext context) {